rename socket.py to listen_socket.py to avoid circular import
